#43:_75%_CRUDGenre

+++ backend/src/main/java/com/nhom4/moviereservation/repository/UserRepository.java
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     Optional<User> findByEmail(String email);
-    Optional<User> findById(Long id);
-    void deleteById(Long id);
+    Optional<User> findById(Integer id);
+    void deleteById(Integer id);
 }
